fix(server): redact tool input on the SDK permission broadcast path (#6038)
The hook path (ws-permissions.js) was sanitized in #6029/#6037, but the SDK/TUI
provider path (permission-manager.js — the DEFAULT provider) built its broadcast
permission_request with input: toolInput verbatim and a description from raw
JSON.stringify(toolInput). session-manager forwards permission_request to every
subscribed client, so a secret in a tool-input value leaked — the same class
#6029 fixed, on the more commonly-exercised path.

- Relocate sanitizeToolInput + redactDeep (+ the depth/size constants) from
  ws-permissions.js into redaction.js (the leaf SSoT). permission-manager.js can
  now import them without pulling in the HTTP-handler stack or risking a cycle.
  ws-permissions.js imports + re-exports sanitizeToolInput (back-compat).
- permission-manager.js: both permPayload sites (tool permission + MCP spawn)
  now broadcast description: redactValue(...) and input: sanitizeToolInput(...).
  The raw input for execution stays on the _pendingPermissions entry, untouched.
- Test: secret-in-value, secret-nested-in-object, and raw-input-preserved.

Validated in a Linux container: new test 3/3, existing redaction test 21/21,
permission-manager 83/83.
